perf(bench): isolate CKF benchmark feature graphs
