+++ lib/bench/Cargo.toml
@@ -17,6 +17,8 @@ description = "Lightweight HTTP benchmarks for Dynamo endpoints"
 default = ["claude-trace-export", "test-support"]
 active-sequences = [
   "router-bench",
+  "dep:dynamo-mocker",
+  "dep:indicatif",
   "dep:dynamo-tokens",
   "dep:libc",
   "dep:rustc-hash",
@@ -31,15 +33,37 @@ claude-trace-export = [
   "dep:sha2",
   "dep:tokenizers",
 ]
-dc-ckf-bench = []
-mooncake = ["router-bench", "dep:libc", "dep:tokio-util"]
-multiturn = ["dep:futures-util", "dep:rand", "dep:reqwest"]
-offline-replay = []
+dc-ckf-consumer = []
+dc-ckf-relay = [
+  "dep:crossbeam-channel",
+  "dep:dynamo-mocker",
+  "dep:indicatif",
+  "dep:libc",
+  "dep:rustc-hash",
+  "dep:sha2",
+]
+mooncake = [
+  "router-bench",
+  "dep:dynamo-mocker",
+  "dep:indicatif",
+  "dep:libc",
+  "dep:tokio-util",
+]
+multiturn = [
+  "dep:dynamo-mocker",
+  "dep:futures-util",
+  "dep:indicatif",
+  "dep:rand",
+  "dep:reqwest",
+]
+offline-replay = ["dep:dynamo-mocker"]
 router-bench = ["dynamo-kv-router/bench"]
 router-metrics = ["dynamo-kv-router/metrics"]
 router-test-support = [
   "router-bench",
   "router-metrics",
+  "dep:dynamo-mocker",
+  "dep:indicatif",
   "dep:dynamo-tokens",
   "dep:libc",
   "dep:rustc-hash",
@@ -88,7 +112,7 @@ required-features = ["mooncake"]
 [[bin]]
 name = "dc_ckf_consumer_bench"
 path = "kv_router/dc_ckf_consumer_bench.rs"
-required-features = ["dc-ckf-bench"]
+required-features = ["dc-ckf-consumer"]
 
 [[bench]]
 name = "active_sequences_bench"
@@ -100,7 +124,7 @@ required-features = ["active-sequences"]
 name = "dc_ckf_relay_bench"
 path = "kv_router/dc_ckf_relay_bench.rs"
 harness = false
-required-features = ["dc-ckf-bench"]
+required-features = ["dc-ckf-relay"]
 
 [dependencies]
 anyhow = { workspace = true }
@@ -110,7 +134,7 @@ crossbeam-channel = { version = "0.5", optional = true }
 flate2 = { version = "1", optional = true }
 futures-util = { workspace = true, optional = true }
 hf-hub = { workspace = true, features = ["ureq"], optional = true }
-indicatif = "0.18"
+indicatif = { version = "0.18", optional = true }
 rand = { workspace = true, optional = true }
 reqwest = { workspace = true, optional = true }
 rustc-hash = { version = "2", optional = true }
@@ -128,7 +152,7 @@ tracing-subscriber = { workspace = true, optional = true }
 uuid = { workspace = true, optional = true }
 dynamo-tokens = { workspace = true, optional = true }
 dynamo-data-gen = { workspace = true, optional = true }
-dynamo-mocker = { workspace = true }
+dynamo-mocker = { workspace = true, optional = true }
 dynamo-kv-router = { workspace = true }
 
 [target.'cfg(target_os = "linux")'.dependencies]
