+++ lib/bench/kv_router/common/mod.rs
@@ -1,12 +1,51 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(any(
+    feature = "active-sequences",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod args;
-#[cfg(feature = "dc-ckf-bench")]
+#[cfg(any(feature = "dc-ckf-consumer", feature = "dc-ckf-relay"))]
+pub mod dc_ckf_metadata;
+#[cfg(feature = "dc-ckf-relay")]
 pub mod dc_ckf_shared;
+#[cfg(any(
+    feature = "dc-ckf-relay",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod issuer;
+#[cfg(any(
+    feature = "active-sequences",
+    feature = "dc-ckf-relay",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod progress;
+#[cfg(any(
+    feature = "active-sequences",
+    feature = "dc-ckf-relay",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod replay;
+#[cfg(any(
+    feature = "active-sequences",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod results;
+#[cfg(any(
+    feature = "active-sequences",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod sweep;
+#[cfg(any(
+    feature = "active-sequences",
+    feature = "mooncake",
+    feature = "router-test-support"
+))]
 pub mod trace_gen;
