+++ lib/bench/kv_router/dc_ckf_consumer_bench.rs
@@ -20,7 +20,7 @@ use std::time::{Duration, Instant};
 
 use anyhow::{Context, ensure};
 use clap::{Parser, ValueEnum};
-use dynamo_bench::kv_router_common::dc_ckf_shared::DcCkfCorpusMetadata;
+use dynamo_bench::kv_router_common::dc_ckf_metadata::DcCkfCorpusMetadata;
 use dynamo_kv_router::identity::{
     CacheSemanticsId, DcId, IdentitySource, IndexerDomainId, PoolId, RoutingScopeId,
 };
